@@ -5,13 +5,12 @@
 // or any plugin's vendor/assets/javascripts directory can be referenced here using a relative path.
 //
 // It's not advisable to add code directly here, but if you do, it'll appear at the bottom of the
-// compiled file.
+// compiled file. JavaScript code in this file should be added after the last require_* statement.
 //
 // Read Sprockets README (https://github.com/rails/sprockets#sprockets-directives) for details
 // about supported directives.
 //= require jquery
 //= require pickadate/picker
 //= require pickadate/picker.date
-//= require bootstrap-sprockets
 //= require topsekrit
 //= require ui-functions
