@@ -12,9 +12,15 @@ Rails.application.configure do
   # preloads Rails for running tests, you may have to set it to true.
   config.eager_load = false
 
-  # Configure static file server for tests with Cache-Control for performance.
-  config.serve_static_files   = true
-  config.static_cache_control = 'public, max-age=3600'
+  # Enable CSRF Tokens per form
+  config.action_controller.per_form_csrf_tokens = true
+  config.action_controller.forgery_protection_origin_check = true
+
+  # Configure public file server for tests with Cache-Control for performance.
+  config.public_file_server.enabled = true
+  config.public_file_server.headers = {
+    'Cache-Control' => 'public, max-age=3600'
+  }
 
   # Show full error reports and disable caching.
   config.consider_all_requests_local       = true
@@ -25,6 +31,8 @@ Rails.application.configure do
 
   # Disable request forgery protection in test environment.
   config.action_controller.allow_forgery_protection = false
+  config.action_mailer.perform_caching = true
+
 
   # Tell Action Mailer not to deliver emails to the real world.
   # The :test delivery method accumulates sent emails in the
@@ -32,9 +40,6 @@ Rails.application.configure do
   config.action_mailer.delivery_method = :test
   config.action_mailer.default_options = {from: 'SecretLink.org Sharing <info@SecretLink.org>'}
   config.action_mailer.default_url_options = { :host => "localhost", :port => 3000 }
-
-  # Randomize the order test cases are executed.
-  config.active_support.test_order = :random
 
   # Print deprecation notices to the stderr.
   config.active_support.deprecation = :stderr
